+++ app/services/transform/execute.py
@@ -6,9 +6,9 @@ from sdx_gcp.app import get_logger
 from app.definitions.executor import ExecutorBase
 from app.definitions.spec import ParseTree, Transform, BuildSpecError, Template, Transforms
 from app.definitions.data import Value, Field, Data
-from app.transform.interpolate import interpolate
-from app.transform.populate import populate_mappings
-from app.transform.tree_walker import TreeWalker
+from app.services.transform.interpolate import interpolate
+from app.services.transform.populate import populate_mappings
+from app.services.transform.tree_walker import TreeWalker
 
 logger = get_logger()
