+++ app/services/transform/interpolate.py
@@ -5,7 +5,7 @@ from sdx_gcp.app import get_logger
 
 from app.definitions.spec import Template, Transforms, ParseTree, BuildSpecError
 from app.definitions.data import Field
-from app.transform.tree_walker import TreeWalker
+from app.services.transform.tree_walker import TreeWalker
 
 logger = get_logger()
