+++ app/services/transform/populate.py
@@ -5,7 +5,7 @@ from sdx_gcp.app import get_logger
 
 from app.definitions.spec import ParseTree
 from app.definitions.data import Data, Field
-from app.transform.tree_walker import TreeWalker
+from app.services.transform.tree_walker import TreeWalker
 
 logger = get_logger()
