+++ tests/unit/transform/functions/test_list.py
@@ -1,7 +1,7 @@
 import unittest
 
 from app.definitions.data import Empty
-from app.transform.functions.lists import as_list, append_to_list, prepend_to_list, trim_list
+from app.services.transform.functions.lists import as_list, append_to_list, prepend_to_list, trim_list
 
 
 class ListTests(unittest.TestCase):
