+++ tests/unit/transform/functions/test_numerical.py
@@ -1,7 +1,7 @@
 import unittest
 
 from app.definitions.data import Empty
-from app.transform.functions.numerical import round_half_up, aggregate, mean, number_equals, total
+from app.services.transform.functions.numerical import round_half_up, aggregate, mean, number_equals, total
 
 
 class RoundTests(unittest.TestCase):
