+++ tests/unit/transform/test_tree_walker.py
@@ -2,7 +2,7 @@ import unittest
 
 from app.definitions.spec import ParseTree
 from app.definitions.data import Field
-from app.transform.tree_walker import TreeWalker
+from app.services.transform.tree_walker import TreeWalker
 
 
 class TreeWalkerTests(unittest.TestCase):
